+++ acme/Configuration.py
@@ -35,6 +35,7 @@ class Configuration(object):
 		argsImportDirectory		= args.importdirectory if args is not None and 'importdirectory' in args else None
 		argsAppsEnabled			= args.appsenabled if args is not None and 'appsenabled' in args else None
 		argsRemoteCSEEnabled	= args.remotecseenabled if args is not None and 'remotecseenabled' in args else None
+		argsValidationEnabled	= args.validationenabled if args is not None and 'validationenabled' in args else None
 
 
 		config = configparser.ConfigParser(	interpolation=configparser.ExtendedInterpolation(),
@@ -92,6 +93,7 @@ class Configuration(object):
 				'cse.enableNotifications'			: config.getboolean('cse', 'enableNotifications', 		fallback=True),
 				'cse.enableRemoteCSE'				: config.getboolean('cse', 'enableRemoteCSE', 			fallback=True),
 				'cse.enableTransitRequests'			: config.getboolean('cse', 'enableTransitRequests',		fallback=True),
+				'cse.enableValidation'				: config.getboolean('cse', 'enableValidation', 			fallback=True),
 				'cse.sortDiscoveredResources'		: config.getboolean('cse', 'sortDiscoveredResources',	fallback=True),
 				'cse.checkExpirationsInterval'		: config.getint('cse', 'checkExpirationsInterval',		fallback=60),		# Seconds
 
@@ -233,6 +235,10 @@ class Configuration(object):
 		# Override remote CSE enablement
 		if argsRemoteCSEEnabled is not None:
 			Configuration._configuration['cse.enableRemoteCSE'] = argsRemoteCSEEnabled
+
+		# Override validation enablement
+		if argsValidationEnabled is not None:
+			Configuration._configuration['cse.enableValidation'] = argsValidationEnabled
 
 		# Correct urls
 		Configuration._configuration['cse.remote.address'] = Utils.normalizeURL(Configuration._configuration['cse.remote.address'])