+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.5.0 - xxxx-xx-xx
 - [CSE] Added support for 'm2m:dbg' throughout the CSE.
+- [CSE] Added configuration and command line argument to enable and disable validation of attributes and arguments.
 - [TESTS] Added test cases for unit tests.